fix(codemagic_app_preview): fix posting comment on GitLab (#111)
The Gitlab comment was getting cut off on the first `&` symbol in the
query param of the `qrCodeUrl`.

<img width="890" alt="Screenshot 2024-12-20 at 14 58 17"
src="https://github.com/user-attachments/assets/0ca6daff-82d8-41ab-ba7a-cea119ad6e14"
/>

This was happening since the Gitlab API takes in the body of the comment
as a query param as well, so the api was parsing the `&` as another
query param on the endpoint call.
The proposed fix encodes the comment body first before calling the API.

---------

Co-authored-by: Andre Ranito <andre.ranito@knowunity.com>
Co-authored-by: Nils Reichardt <me@nils.re>